# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: eSPX Management JSON API
  version: 1.0.0
  description: |
    Machine-readable contract for `/api/v1` JSON routes served by `cmd/management`
    (`internal/management` + `internal/adminapi`). HTMX `/admin/*` HTML routes are excluded.
    HTTP 501 stub routes (deferred dashboards/reports) are excluded — see `internal/adminapi/stub_routes.go` and `internal/openapi/stubs.go`.
servers:
  - url: http://127.0.0.1:8188
    description: Local management (default SERVER_PORT)
components:
  securitySchemes:
    AdminAPIKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: Shared admin API key (`ADMIN_API_KEY`).
    SessionCookie:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT session cookie set by `POST /api/v1/auth/login`.
    ConsentHMAC:
      type: apiKey
      in: header
      name: X-Consent-Signature
      description: HMAC-SHA256 of raw request body (`CONSENT_HMAC_SECRET`).
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    GenericObject:
      type: object
      additionalProperties: true
...
paths:
...
  /api/v1/billing/invariant:
    get:
      tags: [Billing]
      summary: GET /api/v1/billing/invariant
      operationId: get_billing_invariant
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/invoices:
    get:
      tags: [Billing]
      summary: GET /api/v1/billing/invoices
      operationId: get_billing_invoices
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/invoices/preview:
    post:
      tags: [Billing]
      summary: POST /api/v1/billing/invoices/preview
      operationId: post_billing_invoices_preview
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/invoices/{id}:
    get:
      tags: [Billing]
      summary: GET /api/v1/billing/invoices/{id}
      operationId: get_billing_invoices_id
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/billing/invoices/{id}/ledger-lines:
    get:
      tags: [Billing]
      summary: GET /api/v1/billing/invoices/{id}/ledger-lines
      operationId: get_billing_invoices_id_ledger_lines
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/invoices/{id}/pdf:
    get:
      tags: [Billing]
      summary: GET /api/v1/billing/invoices/{id}/pdf
      operationId: get_billing_invoices_id_pdf
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: PDF document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/invoices/{id}/void:
    post:
      tags: [Billing]
      summary: POST /api/v1/billing/invoices/{id}/void
      operationId: post_billing_invoices_id_void
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/summary:
    get:
      tags: [Billing]
      summary: GET /api/v1/billing/summary
      operationId: get_billing_summary
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/customers/{id}/billing/forecast:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/billing/forecast
      operationId: get_customers_id_billing_forecast
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/{id}/billing/statement:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/billing/statement
      operationId: get_customers_id_billing_statement
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/customers/{id}/tax-profile:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/tax-profile
      operationId: get_customers_id_tax_profile
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Customers]
      summary: PUT /api/v1/customers/{id}/tax-profile
      operationId: put_customers_id_tax_profile
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/customers/{id}/wallet:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/wallet
      operationId: get_customers_id_wallet
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/selfserve/billing/statement:
    get:
      tags: [Customers]
      summary: GET /api/v1/selfserve/billing/statement
      operationId: get_selfserve_billing_statement
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...